+++ src/font-info.h
@@ -9,7 +9,7 @@ struct font_metrics {
 };
 
 int get_font_info(const char* font_path,
-                  int font_size,
+                  double font_size,
                   struct font_metrics* metrics_out);
