+++ src/harfbuzz.h
@@ -6,7 +6,7 @@
 
 
 int init_font(const char* font_path,
-              int font_size,
+              double font_size,
               hb_font_t** font_out);
 
 int init_buffer(const char* string,
