+++ src/string-info.c
@@ -4,7 +4,7 @@
 
 int calc_string_width(const char* string,
                       const char* font_path,
-                      int font_size,
+                      double font_size,
                       double* width_out) {
   int error = 0;
 
@@ -35,7 +35,7 @@ int calc_string_width(const char* string,
 
 int calc_string_info(const char* string,
                      const char* font_path,
-                     int font_size,
+                     double font_size,
                      struct string_metrics* metrics_out) {
   int error = 0;
